requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8
detectron2 @ git+https://github.com/facebookresearch/detectron2.git  # Faster R-CNN

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# Dataset Handling
pycocotools>=2.0.6
albumentations>=1.3.0

# Metrics and Evaluation
scikit-learn>=1.3.0
scipy>=1.11.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
tqdm>=4.65.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
jupyter>=1.0.0
ipywidgets>=8.0.0

# Report Generation (optional)
reportlab>=4.0.0
jinja2>=3.1.0
